Update patch set 5
Patch Set 5:

[Replicating this comment because Gerrit seems to have messed up the formatting. Ignore the previous version. Also I missed a part of your reply and added another response.]


 > > So, the goal would be to take the SSL state machine's internal
 > > record-layer buffer and pass it start to the socket to get the
 > > ciphertext.
 > 
 > In that case the socket must be allowed to take owner ship of the
 > buffer though? In some cases (windows) the the socket might outlive
 > SSLSocket, and deleting the buffer will cause problems. That's why
 > I added the BIO_new_bio_pair_external_buf function.
 

Hrm. I'm not very familiar with Windows, but is that true for Chromium's sockets at least. The API contract for net::Socket is that the buffer passed to Read must be alive until either the operation completes OR the socket is disconnected.

It doesn't appear to even be implemented in terms of Windows' overlapped I/O logic, but if it were, wouldn't CancelIo work? We just need to ensure that the net::Socket <-> BIO adapter logic disconnects the socket at the appropriate point and that, likewise, SSL calls the analogous function on its BIO.


 > (For the application's buffers, ciphertext -> plaintext
 > > is going to necessarily involve a copy. I haven't thought about
 > > plaintext -> application buffer yet. You'd have to decrypt the
 > > entire record before returning anything, and the application may
 > > have provided a smaller buffer than the size of the record.)
 > 
 > If the buffer is too small, it should be possible verify first
 > without storing the decrypted text, mark the record as good, and
 > then start decrypting parts of it into the application buffer? The
 > extra cpu usage wouldn't be too bad, if this doesn't happen very
 > often. I would think that normally the caller's buffers are bigger
 > than 16k.


Hrm? I dunno how easily you can verify the MAC without decrypting the whole record first. That's not the API exposed by EVP_AEAD. For legacy cipher suites, they're all MAC-then-encrypt so you'd at least have to be a bit clevr about it.
But, yeah, I agree that normally the record will fit in the caller's buffer, so we can allocate a larger buffer for that case as needed.


 
 > So the way the zero copy API works, both the bio pair read and
 > write buffers will be owned by SSLClientSocketOpenSSL. If SSL state
 > machine internally uses the zero copy API to grab the bio pair
 > read/write buffers it should be possible to remove internal record
 > buffers similarly to how you propose.


Wouldn't that require that the SSL state machine know a priori that it's dealing with a BIO pair and not an arbitrary BIO?


 > Your suggestion could simplify the call flow quite a bit though.
 > The way the application has run go back and forth between both ends
 > of BoringSSL is quite cumbersome. However, Chromium seems to have
 > got this right.


To be fair, it only got this right (write error issue aside) after I wrestled quite a bit with transport error plumbing. ;-) Which was part of the motivation of trying to avoid the running back and forth thing. Chromium's logic makes a lot of assumptions about exactly what an SSL_read or SSL_write may block on, and I think it'd be cleaner if that were just captured somehow by a custom BIO which did the right thing based on whether BIO_read or BIO_write got called.


 > > (This is a much more violent change than this CL and isn't
 > > necessarily mutually exclusive. No reason we couldn't do your
 > > version and then later, if this more violent one is worthwhile,
 > > implement that and remove the now unused API.)
 > 
 > Yes, I think zero copy API would be a good thing to get in for now,
 > independently if the async API is implemented later or not. Since
 > this is a purely additional API, it adds little or no complexity to
 > the existing code, and it will currently remove (at most) 2*17K
 > buffers per active SSL socket. Also, minimal changes is needed on
 > the caller code to adopt this (I have some test code running), so
 > the work is already more or less done. It actually makes the async
 > socket code less complex since the read and write call flows now
 > becomes equal.


Yeah, going with the simpler approach first SGTM. If we end up doing a more complicated one, we can always remove this API if it then becomes unused.
I'm curious what the simpler caller logic looks like (also if you have numbers on what the perf gains are). Could you perhaps uploaded it to codereview.chromium.org?

Patch-set: 5
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0