Update patch set 1
Patch Set 1:

(1 comment)

> I don't suppose the tests would catch this sort of thing?

Not really unless it crashed something. Maybe we could write something for Windows that sets all the calle saved registers to something unique, calls a function and then checks that everything is where it should be afterwards? It would have to be written in asm however.

Patch-set: 1
Label: Code-Review=0