Update patch set 1
Patch Set 1: Code-Review+2

This is a little more complicated because it could be reasonable to say that the function only takes ownership if it returns 1. However, having checks instances where this is called in the rest of the code, that appears to be confused and, in non-BoringSSL code, people often don't check the return value.

Patch-set: 1
Label: Code-Review=+2