RFC: SSL_CIPHER_STATE declarations
This is not for landing since it doesn't do anything, but just to get early
sanity checks on plans. This is to split the cipher bits into a self-contained
component. (Prereq to then splitting out the record layer as this is
more-or-less the only thing shared between their record layers.)  As bonus,
we'll get EVP_AEAD + DTLS working for free.

Plans:
- #includes of ssl_locl.h will be renamed to internal.h. Things in ssl_locl.h
  graduate to internal.h as they get:
  - documented
  - dependencies and layering sorted out
  - generally tidied up
  - implementations moved to sensible file with sensible filename
  - header order matches source order
- Possibly do something similar with ssl.h and some ssl_original.h or whatever?
  Or maybe just have a big comment:
   /* The following things still need to be cleaned up. */
- Soft prereq: SSL_CIPHER_STATE depends on SSL_CIPHER so ideally that'll get
  cleaned up and documented first.
- SSL_CIPHER_STATE subsumes tls1_enc, tls1_mac, tls1_setup_key_block, ssl3
  equivs, {s3,d1}_pkt.c code that calls it and hacks around eivlen.
- DTLS rexmit takes ownership of previous epoch's SSL_CIPHER_STATE, so EVP_AEAD
  ownership problem also resolved. (Not that we'll care when DTLS renego goes
  away.)

(Not going to start actually implementing this until the great clang-formatting
is done.)

Change-Id: I7d4fb9eadf131e1d79c86f26d6595a65d1754b36