Update patch set 1
Patch Set 1:

I haven't read enough of the surrounding code to understand why it is const. However, removing its constness will result in needing to make changes elsewhere - it is passed into functions that expect a const SSL_CIPHER **, and a SSL_CIPHER ** can't be implicitly converted to const.

I understand why a const char * can't be freed (free modifies the memory pointed to, in this example const char), but I don't understand why msvc is complaining about freeing a const SSL_CIPHER **: the memory being freed is type const SSL_CIPHER *, which is not a const type (it's a pointer to a const, but the pointer isn't const).

Patch-set: 1
Reviewer: Gerrit User 5440 <5440@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0