Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

NACK on this temporarily.

I believe this crash doesn't affect us because we don't use DTLSv1_listen. (And it's gone.) The description from upstream is slightly misleading because listen is less a state and more a separate API mode. Upstream's follow-on commit confirms this.
https://git.openssl.org/gitweb/?p=openssl.git;a=commit;h=248385c606620b29ecc96ca9d3603463f879652b

That code and feature don't exist anymore. (It and server-side cookie handling in general has some pretty serious problems. We can probably nuke it.) With that removed, all this CL does is reject DTLS records where the record header and body do not come in the same read.

This is very sensible and something we definitely want... but it breaks WebRTC. Fortunately, the segfault isn't an issue for us, so we can hold on this until we sort out WebRTC. I'll file a bug for them.

Patch-set: 1
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=-1