Update patch set 1
Patch Set 1:

> I assumed "Fix build" is going to be enough, but I should have been
 > more descriptive, sorry about that.
 > 
 > Without those changes, build is broken on OpenBSD with gcc because
 > of the -Wshadow, that you've added a while ago.

Really? What GCC version? We're building with GCC 4.9 in places and -Wshadow doesn't cause those issues.

Since this issue doesn't occur elsewhere, if we accepted it as a convention then we would likely be forever breaking it anyway because of the lack of automated checks. Thus the better path forward might be to remove -Wshadow on OpenBSD if CMake can figure that out.

Patch-set: 1