Make |key| a mandatory parameter to HMAC_Init_ex().
Before you could pass NULL, and this would mean to re-use the existing
key from the HMAC context.

This API contract was easy to misuse, since callers generically passing
in a key as (buffer, length) might accidentally invoke this behavior
for empty keys, where the length would be zero and buffer might be NULL.
(An empty HMAC key is in practice equivalent to an all zero key)

To make things less mysterious, and resolve a bug in PBKDF2, a NULL key
is no longer special.

This change also deletes HMAC_Init(), as it was slated for deprecation
and only used in one place.

BUG=http://crbug.com/449409

Change-Id: I6c4fe6f8d639cb2d7dcf472992114b5f647b033a